final/src/requirements.txt
IMDbPY==2021.4.18
pandas==1.3.4
tmdbv3api==1.7.6

